requirements.txt
# Core Testing Frameworks
# Web Demo Dependencies
flask==2.3.3
flask-socketio==5.3.6
eventlet==0.33.3
python-socketio==5.8.0
playwright==1.40.0
pytest==7.4.3
pytest-html==4.1.1
pytest-json-report==1.5.0
requests==2.31.0
appium-python-client==3.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Report Generation
jinja2==3.1.2
weasyprint==60.2
reportlab==4.0.7

# Storage and Cloud
boto3==1.34.0
minio==7.2.0

# Security Testing
python-owasp-zap-v2.4==0.0.21

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
click==8.1.7
colorama==0.4.6

# Orchestrator Dependencies
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiofiles==23.2.1
watchdog==3.0.0
croniter==2.0.1
psutil==5.9.6
asyncio-mqtt==0.16.1

# Monitoring and Metrics
prometheus-client==0.19.0
grafana-api==1.0.3

# Additional Web Framework
flask==3.0.0
flask-socketio==5.3.6
aiohttp==3.9.1
aiohttp-jinja2==1.5.1

# Configuration Management
pyyaml==6.0.1
jsonschema==4.20.0

# Phase 5 - Business & Collaboration Features
# Data Analysis and Visualization
pandas==2.1.4
numpy==1.25.2
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# AI and Machine Learning
openai==1.3.7
anthropic==0.7.8
scikit-learn==1.3.2
transformers==4.36.2

# Business Intelligence and Analytics
dash==2.14.2
dash-bootstrap-components==1.5.0
streamlit==1.28.2

# Communication and Collaboration
slack-sdk==3.26.1
pymsteams==0.2.2
discord-webhook==1.3.0

# Voice and Audio Processing
pyttsx3==2.90
speechrecognition==3.10.0
gtts==2.4.0
pygame==2.5.2

# Advanced Data Processing
scipy==1.11.4
statsmodels==0.14.1
openpyxl==3.1.2
xlsxwriter==3.1.9

# Time Series Analysis
prophet==1.1.5
pmdarima==2.0.4

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Image and Chart Generation
pillow==10.1.0
kaleido==0.2.1

# Email Integration
smtplib2==0.2.1
email-validator==2.1.0

# Additional Utilities for Business Features
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
cachetools==5.3.2